Refactor demo into modules and add smoke test
