+++ ironcortex/config.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+
+@dataclass
+class CortexConfig:
+    R: int = 32
+    d: int = 128
+    V: int = 256           # vocab size (e.g., bytes)
+    K_inner: int = 8
+    B_br: int = 2
+    k_active: int = 8
+    max_T: int = 8192
